+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update `ruff` to `0.11.9` <https://github.com/gtronset/beets-filetote/pull/183>
 - Provide `filetote:default` to configure custom default renaming for otherwise
   unspecified artifacts <https://github.com/gtronset/beets-filetote/pull/181>
+- Import `annotations` and improve typing for py38+ <https://github.com/gtronset/beets-filetote/pull/184>
 
 ### Fixed